+++ xrCore/_stl_extensions.h
@@ -92,7 +92,7 @@ struct xr_allocator {
 		typedef xalloc<T>	result;
 	};
 
-	static	void	*alloc		(const size_t &n)	{	return xr_malloc(n);	}
+	static	void	*alloc		(const u32 &n)	{	return xr_malloc(n);	}
 	template <typename T>
 	static	void	dealloc		(T *&p)			{	xr_free(p);					}
 };
@@ -122,7 +122,7 @@ public:
 			xr_vector			()									: inherited	()					{}
 			xr_vector			(size_t _count, const T& _value)	: inherited	(_count,_value)		{}
 	explicit xr_vector			(size_t _count)						: inherited (_count)			{}
-	size_t		size				() const							{ return inherited::size();} 
+	u32		size				() const							{ return inherited::size();} 
 
 	void	clear_and_free		()									{ inherited::clear();			}
 	void	clear_not_free		()									{ erase(begin(),end());	}
@@ -132,7 +132,7 @@ public:
 		if ( capacity() <= (size()+size()/4) ) clear_not_free(); 
 		else 
 		{ 
-			size_t old=size(); 
+			u32 old=size(); 
 			clear_and_free(); 
 			reserve(old); 
 		}
