+++ xrCore/_vector2.h
@@ -82,7 +82,7 @@ public:
 		return *this;	
 	}
 
-    T& operator[] (size_t i) const
+    T& operator[] (int i) const
     {
         // assert:  0 <= i < 2; x and y are packed into 2*sizeof(float) bytes
         return (T&) *(&x + i);
